Creating AuthSignUpRequest
